backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
openai
